Changes interface for planning tasks
Removes operation for absolute and relative moves; operation triggerPathPlanning remains for test scripts. This is done due to the architecture update of the AreaExploration Task. Planning will now start when new data is detected in the input ports or through the operation triggerPathPlanning. In the latter, the start and goal pose will be taken from the arguments of the operation.